All merge conflicts solved
Ah, mister Visual.
Mister Github, sir.
Did'ya hear the news about good old Integrator Project
No
You know the X-Gam team?
Yeah
They are trying to do a commit, the physics legacy is secure
Sure
And all they had to do was commit
That's a whole lot of work
If they want a good mark
Ha
Now how're you gonna fuck them up this time
I guess I'm gonna fin'ly have to hardcorly bug
Really
Talk less, implode more
Ha
Do whatever it takes to get this commit to the other world
Now, Adri and Aaron are merciless
Well, hate the program, love the programmer
Git
I'm sorry Visual, I've gotta go
But
Mental nervious crisis are happening over dinner

Two code software and an idiot existing in the same PC
Diametric'ly opposed, foes
They emerge with a compromise, having opened doors that were
Previously closed
So
The idiot emerges with the capability of doing any exe of his whim
A program he can shape however he wants
The softwares emerge with the projects control
And here's the pièce de résistance
No one else was in
The Git where it happened
The Git where it happened
The Git where it happened
No one else was in
The Git where it happened
The Git where it happened
The Git where it happened
No one really knows how the game is played
The art of the commit
How the merge gets made
We just assume that it happens
But no one else is in
The Git where it happens
Visual claims
Team Explorer asked me to make a commit
In distress 'n disarray
Visual claims
The idiot said
I've nothing else to change
And basic'ly begged me to join the github project
Visual claims
I approached Github and said
I know you hate 'im, but let's hear what he has to push
Visual claims
Well, I arranged the meeting
I arranged the menu, the venue, the seating
But
No one else was in
The Git where it happened
The Git where it happened
The Git where it happened
No one else was in
The Git where it happened
The Git where it happened
The Git where it happened
No one really knows how the
fucking Linker goes
The lines of code sacrificed in
Ev'ry commit that's made
We just assume that it happens
But no one else is in
The Git where it happens
Meanwhile
Github is grappling with the fact that not ev'ry issue can be settled by reverting
Meanwhile
Matlab is still fighting over how it even is a thing
It isn't pretty
Then Visual approaches with a push and invite
And Github responds with virtual insight
Maybe we can solve one problem with another and fuck him up so hard he stops programming, in other words
Oh ho
A quid pro quo
I suppose
Wouldn't you like to stop working in this wreck of a home
Actually, I would
Well, I propose the Potomac
And you'll provide him his commit
Well, we'll see how it goes
Let's go
No
One else was in
The Git where it happened
The Git where it happened
The Git where it happened
No one else was in
The Git where it happened
The Git where it happened
The Git where it happened
My God
In God we trust, but we never really knew what got pushed up
Click boom and it happened
But no one else was in the Git where it happened.
You idiotic moron
What happened here to convince you that you deserved happiness?
You retarded knucklehead
Where you thinking about going to have dinner
Was there hunger pressure to deliver
Wanky wanker piece of shit
Or did you know, even then, it doesn't matter
From where you even try to commit
'Cause we'll have the lines
We're in the best spot
You got more than you gave
And I wanted what I got
When you got skin in the game, you stay in the game
But you don't get a win unless you play in the game
Oh, you get love for it, you get hate for it
You get nothing if you
Wait for it, wait for it, wait
God help and forgive me
I wanna commit
Something that's gonna
compile decently
What do you want, Git
What do you want, Git
If you don't stand for us
Git, then who do you work for
I
Wanna be in
The Git where it happens
The Git where it happens
I
Wanna be in
The Git where it happens
The Git where it happens
I
Wanna be
In the Git where it happens
I
I wanna be in the Git
Oh
Oh
I wanna be in
The Git where it happens
The Git where it happens
The Git where it happens
I wanna be in the Git
Where it happens
The Git where it happens
The Git where it happen
The art of the online nonsense
Hold your breath and close your project
We want our software to save the day
But we don't get a say in what they trade away
We dream of a brand new integrator
But we integrat in the dark for the most part
Dark as a tomb where it happens
I've got to be in
The Git (where it happens)
I've got to be (the Git where it happens)
I've got to be (the Git where it happens)
Oh, I've got to be in
The Git where it happens
I've got to be, I've gotta be, I've gotta be
In the Git
Click Crash